split ViewBy tests to their own file
